fix format string for Visual Studio
The "z" size specifier is not supported by Visual Studio.
Cast to unsigned long and use "l" specifier as the type has
enough precision for this vector.

Acked-by: Christophe Fergeau <cfergeau@redhat.com>

+++ vdagent/desktop_layout.cpp
@@ -147,7 +147,8 @@ void DesktopLayout::set_displays()
             break;
         }
         if (display_id >= _displays.size()) {
-            vd_printf("display_id %lu out of range, #displays %zu" , display_id, _displays.size());
+            vd_printf("display_id %lu out of range, #displays %lu",
+                      display_id, (unsigned long) _displays.size());
             break;
         }
         DisplayMode * mode(_displays.at(display_id));
